medium workable state damage.js and logs_generator are created
